Move ingest into logic.documents
