+++ aleph/logic/documents/__init__.py
@@ -1,10 +1,10 @@
 import logging
 
-from aleph.core import celery
 from aleph.model import Document
 from aleph.index import documents as index
 from aleph.index.documents import index_document_id
-from aleph.analyze import analyze_document
+from aleph.logic.documents.ingest import ingest, ingest_document  # noqa
+from aleph.logic.documents.ingest import process_document  # noqa
 
 log = logging.getLogger(__name__)
 
@@ -19,16 +19,8 @@ def delete_document(document, deleted_at=None):
     for child in document.children:
         # TODO: are we likely to hit recursion limits?
         delete_document(child, deleted_at=deleted_at)
-
     index.delete_document(document.id)
     document.delete(deleted_at=deleted_at)
-
-
-def process_document(document):
-    """Perform post-ingest tasks like analysis and indexing."""
-    analyze_document(document)
-    index.index_document(document)
-    index.index_records(document)
 
 
 def index_documents(collection_id=None):
