+++ aleph/logic/documents/result.py
